refactor(dialect): split reader into mod/eh/tests for the 2000-line cap
The EH constructors pushed dialect.rs to 2153 lines. Splits the invoke/
landingpad construction into dialect/eh.rs and the reader tests into
dialect/tests.rs; mod.rs lands at 1947.
