+++ crates/perry-codegen/src/dialect/mod.rs
@@ -26,6 +26,10 @@ use inkwell::values::{
 };
 use inkwell::{AddressSpace, FloatPredicate, IntPredicate};
 
+mod eh;
+#[cfg(test)]
+mod tests;
+
 /// Create (only) the function declaration for `fn_text`'s define header, so
 /// later-defined functions are callable while earlier bodies are read. The
 /// native path pre-declares every define before reading any body — calls to
@@ -782,114 +786,6 @@ impl<'ctx, 'm> FnReader<'ctx, 'm> {
             inkwell::values::ValueKind::Basic(v) => Ok(Some(v)),
             _ => Ok(None),
         }
-    }
-
-    /// `[%r =] invoke TY @callee(ARGS) to label %CONT unwind label %PAD`
-    /// (#7302). Shares the callsite-typed, call-through-pointer semantics
-    /// of [`call`]; the only additions are the two edges.
-    fn invoke(&mut self, dst: Option<&str>, rest: &str) -> Result<Option<BasicValueEnum<'ctx>>> {
-        let to_pos = rest
-            .rfind(" to label %")
-            .ok_or_else(|| anyhow!("invoke without `to label`"))?;
-        let head = &rest[..to_pos];
-        let edges = &rest[to_pos + " to label %".len()..];
-        let (cont_label, unwind_label) = edges
-            .split_once(" unwind label %")
-            .ok_or_else(|| anyhow!("invoke without `unwind label`"))?;
-        let cont = self.block(cont_label.trim());
-        let pad = self.block(unwind_label.trim());
-
-        let callee_pos = head
-            .find(['@', '%'])
-            .ok_or_else(|| anyhow!("invoke without callee"))?;
-        let sig_str = head[..callee_pos].trim().trim_end_matches('*').trim();
-        let after = &head[callee_pos..];
-        let paren = after
-            .find('(')
-            .ok_or_else(|| anyhow!("invoke missing arg list"))?;
-        let callee = &after[..paren];
-        let close = rmatch_paren(after, paren)?;
-        let args_str = &after[paren + 1..close];
-
-        // `build_indirect_invoke` takes basic values (not metadata enums
-        // like the call path), so collect both shapes once.
-        let mut args: Vec<BasicValueEnum> = Vec::new();
-        let mut arg_types: Vec<inkwell::types::BasicMetadataTypeEnum> = Vec::new();
-        for a in split_top_level(args_str) {
-            let (aty, atok) = ty_and_val(&a)?;
-            let ty = basic_type(self.ctx, aty)?;
-            args.push(self.val(ty, atok)?);
-            arg_types.push(ty.into());
-        }
-
-        let name = dst.map(|d| d.trim_start_matches('%')).unwrap_or("");
-        let fn_ty = indirect_fn_type(self.ctx, sig_str, &arg_types)?;
-        let callee_ptr = if let Some(fname) = callee.strip_prefix('@') {
-            let n = unquote(fname);
-            let f = match self.module.get_function(&n) {
-                Some(f) => f,
-                None if n.starts_with("llvm.") => self.module.add_function(&n, fn_ty, None),
-                None => bail!("invoke of undeclared @{n}"),
-            };
-            f.as_global_value().as_pointer_value()
-        } else {
-            self.val(self.ctx.ptr_type(AddressSpace::default()).into(), callee)?
-                .into_pointer_value()
-        };
-        let site = self
-            .builder
-            .build_indirect_invoke(fn_ty, callee_ptr, &args, cont, pad, name)
-            .map_err(be)?;
-        // An invoke terminates its block; the emitted text continues in
-        // the inline continuation label, which arrives as the next line.
-        match site.try_as_basic_value() {
-            inkwell::values::ValueKind::Basic(v) => {
-                if let Some(d) = dst {
-                    self.vals.insert(d.trim().to_string(), v);
-                }
-                Ok(Some(v))
-            }
-            _ => Ok(None),
-        }
-    }
-
-    /// `%r = landingpad { ptr, i32 } catch ptr null` (#7302). Perry emits
-    /// exactly one shape — a catch-all pad whose `{ptr, i32}` result is
-    /// unused (the thrown value comes from the runtime's rooted TLS slot),
-    /// so anything else is a dialect drift and bails loudly.
-    fn landingpad(&mut self, dst: &str, rest: &str) -> Result<()> {
-        let body = rest.trim();
-        let clause_pos = body
-            .find("catch")
-            .ok_or_else(|| anyhow!("landingpad without a catch clause: {body}"))?;
-        let ty_tok = body[..clause_pos].trim();
-        if ty_tok.replace(' ', "") != "{ptr,i32}" {
-            bail!("unexpected landingpad type `{ty_tok}`");
-        }
-        if body[clause_pos..].replace(' ', "") != "catchptrnull" {
-            bail!("unexpected landingpad clauses `{}`", &body[clause_pos..]);
-        }
-        let pf = self
-            .func
-            .get_personality_function()
-            .ok_or_else(|| anyhow!("landingpad in a function with no personality"))?;
-        let ptr_ty = self.ctx.ptr_type(AddressSpace::default());
-        let exc_ty = self
-            .ctx
-            .struct_type(&[ptr_ty.into(), self.ctx.i32_type().into()], false);
-        let null = ptr_ty.const_null();
-        let v = self
-            .builder
-            .build_landing_pad(
-                exc_ty,
-                pf,
-                &[null.into()],
-                false,
-                dst.trim_start_matches('%'),
-            )
-            .map_err(be)?;
-        self.vals.insert(dst.trim().to_string(), v);
-        Ok(())
     }
 
     fn call_asm(&mut self, rest: &str) -> Result<()> {
@@ -2046,108 +1942,5 @@ fn apply_flags(inst: Option<InstructionValue<'_>>, flags: &[&str]) {
     }
     if fmf != 0 {
         unsafe { llvm_sys::core::LLVMSetFastMathFlags(inst.as_value_ref(), fmf) };
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn split_corpus(text: &str) -> (String, Vec<String>) {
-        let mut skeleton = String::new();
-        let mut fns = Vec::new();
-        let mut cur: Option<String> = None;
-        for line in text.lines() {
-            if line.starts_with("define ") {
-                cur = Some(String::new());
-            }
-            match cur.as_mut() {
-                Some(f) => {
-                    f.push_str(line);
-                    f.push('\n');
-                    if line == "}" {
-                        fns.push(cur.take().unwrap());
-                    }
-                }
-                None => {
-                    skeleton.push_str(line);
-                    skeleton.push('\n');
-                }
-            }
-        }
-        (skeleton, fns)
-    }
-
-    /// Every function in a real perry-emitted corpus file must construct
-    /// natively and pass the LLVM verifier. This is the reader's primary
-    /// gate: a form it cannot express fails here, not in a user build.
-    fn corpus_roundtrip(path: &str) {
-        // The corpora are tracked in-tree alongside this reader, so a missing
-        // file is a broken checkout, not a branch without artifacts. Skipping
-        // would make the reader's primary gate pass vacuously — precisely the
-        // failure mode the Linux bring-up had to rule out by hand.
-        let text = std::fs::read_to_string(path)
-            .unwrap_or_else(|e| panic!("corpus file {path} is not readable: {e}"));
-        let (skeleton, fns) = split_corpus(&text);
-        let ctx = Context::create();
-        let module = crate::inprocess::parse_ir_text(&ctx, &skeleton, "corpus_skel")
-            .expect("skeleton parses");
-        for f in &fns {
-            predeclare_function_from_text(&ctx, &module, f)
-                .unwrap_or_else(|e| panic!("predeclare: {e:#}"));
-        }
-        let mut n = 0usize;
-        for f in &fns {
-            n += add_function_from_text(&ctx, &module, f).unwrap_or_else(|e| panic!("{e:#}"));
-        }
-        assert!(
-            n > 1000,
-            "expected a real corpus, built only {n} instructions"
-        );
-        module
-            .verify()
-            .unwrap_or_else(|e| panic!("verifier rejected native module:\n{}", e.to_string()));
-    }
-
-    #[test]
-    fn corpus_spike() {
-        corpus_roundtrip(concat!(
-            env!("CARGO_MANIFEST_DIR"),
-            "/../../experiments/llvm-inprocess-spike/spike_text.ll"
-        ));
-    }
-
-    /// #7302: a try/catch/finally corpus — invoke edges, landing pads,
-    /// the personality clause on the define, and the inline continuation
-    /// labels an invoke split leaves behind. Without this the reader's EH
-    /// support would be exercised only by the async spike, which has one
-    /// shape (the async rejection boundary) and no nesting.
-    #[test]
-    fn corpus_exception_handling() {
-        let path = concat!(
-            env!("CARGO_MANIFEST_DIR"),
-            "/../../experiments/llvm-inprocess-spike/eh_text.ll"
-        );
-        // Assert the subject is LIVE: a corpus that lost its EH forms would
-        // still round-trip, and would silently stop testing invoke.
-        let text = std::fs::read_to_string(path).expect("eh corpus readable");
-        assert!(
-            text.matches("invoke ").count() >= 20,
-            "eh corpus has no invoke edges left"
-        );
-        assert!(text.contains("landingpad"), "eh corpus has no landing pad");
-        assert!(
-            text.contains("personality ptr @perry_eh_personality"),
-            "eh corpus lost its personality clause"
-        );
-        corpus_roundtrip(path);
-    }
-
-    #[test]
-    fn corpus_batch_kernel() {
-        corpus_roundtrip(concat!(
-            env!("CARGO_MANIFEST_DIR"),
-            "/../../experiments/llvm-inprocess-spike/batch_kernel.ll"
-        ));
     }
 }
